+++ packages/api/lib/controllers/measurementsController.js
@@ -3,13 +3,14 @@
 const
   { BadRequestError, UnsupportedMediaTypeError } = require('restify-errors'),
   { Measurement, Box } = require('@sensebox/opensensemap-api-models'),
-  csvstringify = require('csv-stringify'),
-  { checkContentType } = require('../helpers/apiUtils'),
+  { checkContentType, createDownloadFilename, csvStringifier } = require('../helpers/apiUtils'),
   {
     retrieveParameters,
     validateFromToTimeParams
   } = require('../helpers/userParamHelpers'),
-  handleError = require('../helpers/errorHandler');
+  handleError = require('../helpers/errorHandler'),
+  OutlierTransformer = require('../transformers/outlierTransformer'),
+  jsonstringify = require('stringify-stream');
 
 /**
  * @api {get} /boxes/:senseBoxId/sensors Get latest measurements of a senseBox
@@ -41,6 +42,11 @@ const getLatestMeasurements = function getLatestMeasurements (req, res, next) {
  * @apiParam {Number=1-50} [outlier-window=15] Size of moving window used as base to calculate the outliers.
  */
 
+const jsonLocationReplacer = function jsonLocationReplacer (k, v) {
+  // dont send unnecessary nested location
+  return (k === 'location') ? v.coordinates : v;
+};
+
 /**
  * @api {get} /boxes/:senseBoxId/data/:sensorId?from-date=fromDate&to-date=toDate&download=true&format=json Get the 10000 latest measurements for a sensor
  * @apiDescription Get up to 10000 measurements from a sensor for a specific time frame, parameters `from-date` and `to-date` are optional. If not set, the last 48 hours are used. The maximum time frame is 1 month. If `download=true` `Content-disposition` headers will be set. Allows for JSON or CSV format.
@@ -56,22 +62,35 @@ const getLatestMeasurements = function getLatestMeasurements (req, res, next) {
  * @apiUse SeparatorParam
  */
 const getData = function getData (req, res, next) {
-  const { sensorId, format, download } = req._userParams;
+  const { sensorId, format, download, outliers, outlierWindow, delimiter } = req._userParams;
+  let stringifier;
 
   // IDEA: add geojson point featurecollection format
   if (format === 'csv' || (download === 'true')) {
     res.header('Content-Type', 'text/csv');
     res.header('Content-Disposition', `attachment; filename=${sensorId}.${format}`);
+    stringifier = csvStringifier(['createdAt', 'value'], delimiter);
   } else if (format === 'json') {
     res.header('Content-Type', 'application/json; charset=utf-8');
+    // IDEA: add geojson point featurecollection format
+    stringifier = jsonstringify({ open: '[', close: ']' }, jsonLocationReplacer);
   }
 
-  Measurement.getMeasurementsStream(req._userParams)
+  let measurementsStream = Measurement.getMeasurementsStream(req._userParams)
     .on('error', function (err) {
-      res.end(`Error: ${err.message}`);
+      return handleError(err, next);
+    });
 
-      return next(err);
-    })
+  if (outliers) {
+    measurementsStream = measurementsStream
+      .pipe(new OutlierTransformer({
+        window: Math.trunc(outlierWindow), // only allow integer values
+        replaceOutlier: (outliers === 'replace')
+      }));
+  }
+
+  measurementsStream
+    .pipe(stringifier)
     .pipe(res);
 };
 
@@ -88,9 +107,10 @@ const getData = function getData (req, res, next) {
  * @apiUse BBoxParam
  * @apiUse ExposureFilterParam
  * @apiParam {String=createdAt,value,lat,lon,height,boxId,boxName,exposure,sensorId,phenomenon,unit,sensorType} [columns=createdAt,value,lat,lon] Comma separated list of columns to export.
+ * @apiParam {Boolean=true,false} [download=true] Set the `content-disposition` header to force browsers to download instead of displaying.
  */
 const getDataMulti = function getDataMulti (req, res, next) {
-  const { boxId, bbox, exposure, delimiter, columns, fromDate, toDate, phenomenon } = req._userParams;
+  const { boxId, bbox, exposure, delimiter, columns, fromDate, toDate, phenomenon, download } = req._userParams;
 
   // build query
   const queryParams = {
@@ -108,8 +128,8 @@ const getDataMulti = function getDataMulti (req, res, next) {
   }
 
   // exposure parameter
-  if (req._userParams.exposure) {
-    queryParams['exposure'] = exposure;
+  if (exposure) {
+    queryParams['exposure'] = { '$in': exposure };
   }
 
   Box.findMeasurementsOfBoxesStream({
@@ -117,16 +137,20 @@ const getDataMulti = function getDataMulti (req, res, next) {
     bbox,
     from: fromDate.toDate(),
     to: toDate.toDate(),
-    columns,
-    transformations: { stringifyTimestamps: true }
+    columns
   })
     .then(function (cursor) {
 
       res.header('Content-Type', 'text/csv');
-      const stringifier = csvstringify({ columns, delimiter, header: 1 });
+      if (download === 'true') {
+        res.header('Content-Disposition', `attachment; filename=${createDownloadFilename(req.date(), 'download', [phenomenon, ...columns], 'csv')}`);
+      }
 
       cursor
-        .pipe(stringifier)
+        .on('error', function (err) {
+          return handleError(err, next);
+        })
+        .pipe(csvStringifier(columns, delimiter))
         .pipe(res);
 
     })
@@ -301,11 +325,12 @@ module.exports = {
       { name: 'boxId', aliases: ['senseboxid', 'senseboxids', 'boxid', 'boxids'], dataType: ['id'] },
       { name: 'phenomenon', required: true },
       { predef: 'delimiter' },
-      { name: 'exposure', allowedValues: Box.BOX_VALID_EXPOSURES },
+      { name: 'exposure', allowedValues: Box.BOX_VALID_EXPOSURES, dataType: ['String'] },
       { predef: 'columnsGetDataMulti' },
       { predef: 'bbox' },
       { predef: 'toDate' },
-      { predef: 'fromDate' }
+      { predef: 'fromDate' },
+      { name: 'download', defaultValue: 'true', allowedValues: ['true', 'false'] },
     ]),
     validateFromToTimeParams,
     getDataMulti