+++ packages/models/src/utils.js
@@ -13,34 +13,35 @@ const parseTimestamp = function parseTimestamp (timestamp) {
     .split('T');
 
   if (!timeParts) {
-    const invalidTimestamp = moment.invalid();
-    invalidTimestamp.__inputString = timestamp;
-
-    return invalidTimestamp;
+    return moment.invalid();
   }
 
-  const [year, month, day] = dateParts.split('-');
-  if (!month || !day) {
-    const invalidTimestamp = moment.invalid();
-    invalidTimestamp.__inputString = timestamp;
-
-    return invalidTimestamp;
+  const [year, month, day, ...restDate] = dateParts.split('-');
+  if (restDate.length !== 0
+    || !month
+    || !day
+    || month.length !== 2
+    || day.length !== 2
+    || year.length !== 4) {
+    return moment.invalid();
   }
 
   /* eslint-disable prefer-const */
   let [hour, minute, lastTimePart] = timeParts.split(':');
   /* eslint-enable prefer-const */
   if (!minute || !lastTimePart || !lastTimePart.endsWith('Z')) {
-    const invalidTimestamp = moment.invalid();
-    invalidTimestamp.__inputString = timestamp;
-
-    return invalidTimestamp;
+    return moment.invalid();
   }
   lastTimePart = lastTimePart.slice(0, -1);
+
 
   /* eslint-disable prefer-const */
   let [second, millisecond = '0'] = lastTimePart.split('.');
   /* eslint-enable prefer-const */
+
+  if ([hour, minute, second].some(tp => tp.length !== 2)) {
+    return moment.invalid();
+  }
 
   // values after the dot are interpreted as nanoseconds
   // if there are more than 3 digits
@@ -48,18 +49,19 @@ const parseTimestamp = function parseTimestamp (timestamp) {
     millisecond = parseInt(millisecond, 10) / 1000000;
   }
 
-  return moment.utc([year, month - 1, day, hour, minute, second, millisecond]);
+  return moment.utc([year, month - 1, day, hour, minute, second, millisecond], undefined, true, true);
 };
 
 const parseAndValidateTimestamp = function parseAndValidateTimestamp (
   timestamp
 ) {
+  const input = timestamp;
   if (!moment.isMoment(timestamp)) {
     timestamp = parseTimestamp(timestamp);
   }
 
   if (!timestamp.isValid()) {
-    throw new Error(`Invalid timestamp '${timestamp.creationData().input}'.`);
+    throw new Error(`Invalid timestamp '${input}'.`);
   }
 
   return timestamp;