+++ tests/tests/013-classify-test.js
@@ -10,24 +10,24 @@ const chakram = require('chakram'),
 
 const BASE_URL = `${process.env.OSEM_TEST_BASE_URL}/boxes`;
 
-describe('openSenseMap API Routes: getBoxes', function () {
-  before('add test data without measurements', function () {
-    chakram.post(`${process.env.OSEM_TEST_BASE_URL}/users/register`, { name: 'classifytestuser', email: 'idwtestuser@test.test', password: '12345678' })
+describe('openSenseMap API Routes: getBoxes Classification', function () {
+  before('add test data', function () {
+    return chakram.post(`${process.env.OSEM_TEST_BASE_URL}/users/register`, { name: 'classifytestuser', email: 'classifytestuser@test.test', password: '12345678' })
       .then(function (response) {
         expect(response.body.token).to.exist;
 
         const jwt = response.body.token;
 
         return chakram.all([
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [ 7.6456, 51.9624 ], name: 'noData' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [ 7.6456, 51.9624 ], name: 'active' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [ 7.6456, 51.9624 ], name: 'inactive' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [7.6456, 51.9624], name: 'noData' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [7.6456, 51.9624], name: 'active' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes`, valid_sensebox({ lonlat: [7.6456, 51.9624], name: 'inactive' }), { headers: { 'Authorization': `Bearer ${jwt}` } }),
         ]);
       })
       .then(function (responses) {
         const boxes = responses.map(r => { return { _id: r.body.data._id, sensorid: r.body.data.sensors[0]._id, name: r.body.data.name }; });
 
-        for (const [ i, box ] of boxes.entries()) {
+        for (const [i, box] of boxes.entries()) {
           box.measurements = [];
           if (box.name === 'senseBoxactive') {
             box.measurements.push({
@@ -51,7 +51,7 @@ describe('openSenseMap API Routes: getBoxes', function () {
           }
         }
 
-        const [ first, second, third ] = boxes;
+        const [first, second, third] = boxes;
 
         return chakram.all([
           chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${first._id}/data`, first.measurements, { headers: { 'content-type': 'application/json' } }),
@@ -62,13 +62,13 @@ describe('openSenseMap API Routes: getBoxes', function () {
   });
 
   after('delete user', function () {
-    chakram.post(`${process.env.OSEM_TEST_BASE_URL}/users/sign-in`, { email: 'classifytestuser@test.test', password: '12345678' })
+    return chakram.post(`${process.env.OSEM_TEST_BASE_URL}/users/sign-in`, { email: 'classifytestuser@test.test', password: '12345678' })
       .then(function (response) {
         expect(response.body.token).to.exist;
 
         const jwt = response.body.token;
 
-        chakram.delete(`${process.env.OSEM_TEST_BASE_URL}/users/me`, { password: '12345678' }, { headers: { 'Authorization': `Bearer ${jwt}` } });
+        return chakram.delete(`${process.env.OSEM_TEST_BASE_URL}/users/me`, { password: '12345678' }, { headers: { 'Authorization': `Bearer ${jwt}` } });
       });
   });