+++ packages/models/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @sensebox/opensensemap-api-models Changelog
 
 ## Unreleased
+- Expand .npmignore
+
+## v0.0.7
+- Fix an error where invalid timestamps were displayed as NaN in returned error messages
+- More strict timestamp parsing (With tests)
+- Remove "stream-transform", "simple-statistics", "csv-stringify" and "stringify-stream" depencies
+- Use mongoose built-in `QueryCursor#map` for data transformation
+- do not populate lastMeasurements in Box.findMeasurementsOfBoxesStream
+- Move Measurements outlier computation into api package
 
 ## v0.0.6
 - Added mqtt-osem-integration MQTT service connection